Tests/LibWeb/Text/expected/createBitmap.txt

RangeError: 0 is an invalid value for sw
Bitmap is correctly loaded
